Add sync_file option for dry run

+++ multipatch_analysis/util.py
@@ -1,7 +1,8 @@
+from __future__ import print_function
 import os, sys, time
 
 
-def sync_file(src, dst):
+def sync_file(src, dst, test=False):
     """Safely copy *src* to *dst*, but only if *src* is newer or a different size.
     """
     if os.path.isfile(dst):
@@ -12,14 +13,14 @@ def sync_file(src, dst):
         if up_to_date:
             return "skip"
         
-        safe_copy(src, dst)
+        safe_copy(src, dst, test=test)
         return "update"
     else:
-        safe_copy(src, dst)
+        safe_copy(src, dst, test=test)
         return "copy"
 
 
-def safe_copy(src, dst):
+def safe_copy(src, dst, test=False):
     """Copy a file, but rename the destination file if it already exists.
     
     Also, the destination file is suffixed ".partial" until the copy is complete.
@@ -28,7 +29,8 @@ def safe_copy(src, dst):
     try:
         new_name = None
         print("copy: %s => %s" % (src, dst))
-        chunk_copy(src, tmp_dst)
+        if test is False:
+            chunk_copy(src, tmp_dst)
         if os.path.exists(dst):
             # rename destination file to avoid overwriting
             now = time.strftime('%Y-%m-%d_%H:%M:%S')
@@ -38,15 +40,18 @@ def safe_copy(src, dst):
                 if not os.path.exists(new_name):
                     break
                 i += 1
-            os.rename(dst, new_name)
-        os.rename(tmp_dst, dst)
+            print("rename:", dst, new_name)
+            if test is False:
+                os.rename(dst, new_name)
+        if test is False:
+            os.rename(tmp_dst, dst)
     except Exception:
         # Move dst file back if there was a problem during copy
-        if new_name is not None and os.path.exists(new_name):
+        if test is False and new_name is not None and os.path.exists(new_name):
             os.rename(new_name, dst)
         raise
     finally:
-        if os.path.isfile(tmp_dst):
+        if test is False and os.path.isfile(tmp_dst):
             os.remove(tmp_dst)
